pyproject.toml
[project]
name = "diary-push-bot"
version = "0.1.0"
description = "Send a random today-in-history diary entry by email."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "python-dotenv>=1.0.1,<2",
  "tzdata>=2025.2,<2026",
]

[dependency-groups]
dev = [
  "pytest>=8.3.5,<9",
]

[project.scripts]
diary-push-bot = "diary_push_bot.cli:main"

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/diary_push_bot"]

[tool.pytest.ini_options]
testpaths = ["tests"]
